создаёт и подключает модуль create-comment
создаёт и подключает модуль для создания комментария к фотографиям
