+++ js/main.js
@@ -1,30 +1,6 @@
-import {PHOTOS_URL, PHOTOS_DESCRIPTION, AVATARS_URL, MESSAGES, NAMES} from './data.js';
-import {getRandomInteger, getRandomArrayElement} from './randomizers.js';
-
-// Функция вызвращает случайный комментарий содержащий один или два сообщения. Количество сообщений определяется случайным образом
-const getRandomMessage = function () {
-  if (getRandomInteger(1, 2) === 2) {
-    const firstMessage = getRandomArrayElement(MESSAGES);
-    let secondMessage = getRandomArrayElement(MESSAGES);
-
-    while (firstMessage === secondMessage) {
-      secondMessage = getRandomArrayElement(MESSAGES);
-    }
-
-    return `${firstMessage} ${secondMessage}`;
-  }
-
-  return getRandomArrayElement(MESSAGES);
-};
-
-let idComment = 0;
-
-const createComment = () => ({
-  id: idComment++,
-  avatar: getRandomArrayElement(AVATARS_URL),
-  message: getRandomMessage(),
-  name: getRandomArrayElement(NAMES)
-});
+import {PHOTOS_URL, PHOTOS_DESCRIPTION} from './data.js';
+import {getRandomInteger} from './randomizers.js';
+import {createComment} from './create-comment.js';
 
 const createPhoto = () => {
   const idPhoto = getRandomInteger(1, 25);
